feat: validation data hook
